requirements-teams.txt
# Additional requirements for Microsoft Teams integration

# Bot Framework Core Components
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-adapter-azure==1.0.0

# HTTP Server for Bot Endpoint
aiohttp==3.8.5
aiohttp-cors==0.7.0

# Environment and Configuration
python-dotenv==1.0.0

# Additional Bot Framework Extensions
botframework-connector==4.15.0
msrest==0.7.1

# Authentication and Security
msal==1.24.0
cryptography==41.0.3

# Utilities
asyncio==3.4.3
pydantic==2.3.0

# Development and Testing (optional)
pytest==7.4.0
pytest-asyncio==0.21.1
ngrok==1.1.0

# Existing workforce management requirements
# (Include all your existing requirements.txt here)
chainlit>=1.0.0
chromadb>=0.4.0
plotly>=5.0.0
kaleido>=0.2.1
autogen>=0.2.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0 